pyproject.toml
[project]
name = "my-claude-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.12.1",
    "claude-agent-sdk>=0.1.50",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "ta>=0.11.0",
    "yfinance>=1.2.0",
]
